Add support for integral fitting
I’ve added a function to integrate columns and rows of the image. I
also modified a few small bits of code make the GUI work better with
the various fitting methods. There is now an integral thermal fit type
which performs gaussian fits on the integrated columns/rows. I have not
updated the cloudpars functions to work properly with this fit type
yet, so the numbers extracted are not correct yet.

DSB