Updated slice residuals for 2D fits
The slice residuals for 2D fits are now averages of the residual over
the slice width. This is done so that the residual shown in the slice
graph corresponds to the deviation of the fit from the slice.
Previously, the 2D fit residuals were always displayed for a slice width
of 1.

DSB